+++ Stocks.DividendUpdater/IUpdater.cs
@@ -1,0 +1,4 @@
+﻿public interface IUpdater
+{
+    Task Update();
+}
